+++ main.c
@@ -20,24 +20,36 @@ typedef enum targ_e {
   ARG_RESERVED = 0x8000, // NOTE: exclusive options below
   ARG_MASK,
   ARG_VNUM,
+  ARG_SCALE,
   ARG_BMP
 } targ_t;
-#define NUM_ARGS 7
+#define NUM_ARGS 8
+static const char* args_[NUM_ARGS] = {
+  "--nocopy", "--verbose", "--raw", "--noinline",
+  "-m", "-v", "-s", "-B"
+};
+static const targ_t arge_[NUM_ARGS] = {
+  ARG_NOCOPY, ARG_VERBOSE, ARG_RAW, ARG_NOINLINE,
+  ARG_MASK, ARG_VNUM, ARG_SCALE, ARG_BMP
+};
 
 static __inline__ int
-print_help_(const char* restrict cmdln)
+phelp_(const char* restrict cmdln)
 {
   fprintf(stderr,
     "Usage: %s [OPTIONS] <data to encode>" __nl
     "OPTIONS:" __nl
-    "\t--nocopy     do not print copyright header" __nl
-    "\t--verbose    print runtime information for generated values" __nl
-    "\t--raw        print generated matrix as 1's and 0's (no Unicode)" __nl
-    "\t--noinline   do not print any inline code, disregards --raw" __nl
-    "\t--mask <N>   force N mask output, regardless of penalty; N:(0-7)" __nl
-    "\t--vnum <N>   tries to force use of N version QR Codes; N:(1-%d)" __nl
-    "\t--bmp <STR>  create STR bitmap file with generated code" __nl,
-    cmdln, MAX_VERSION);
+    "\t--nocopy    do not print copyright header" __nl
+    "\t--verbose   print runtime information for generated values" __nl
+    "\t--raw       print generated matrix as 1's and 0's (no Unicode)" __nl
+    "\t--noinline  do not print any inline code, disregards --raw" __nl
+    "\t-m <N>      force N mask output, regardless of penalty; N:(0-7)" __nl
+    "\t-v <N>      tries to force use of N version QR Codes; N:(1-"
+    __xstr(MAX_VERSION) ")" __nl
+    "\t-s <N>      scale image output by N times; N:(1-"
+    __xstr(MAX_SCALE) ")" __nl
+    "\t-B <STR>    create STR bitmap file with generated code" __nl,
+    cmdln);
   return EINVAL;
 }
 
@@ -46,16 +58,17 @@ main(int argc, char* argv[])
 {
 #ifdef _WIN32
   // NOTE: to allow box-drawing characters
-  system("chcp 65001>nul");
+  system("chcp 65001");
 #endif
 
   if (argc < NUM_MANDATORY + 1)
   {
-    return print_help_(argv[0]);
+    return phelp_(argv[0]);
   }
   targ_t options = ARG_NONE;
   int mask = -1;
   int vnum = -1;
+  int scale = -1;
   int argcount = 0;
   imgfmt_t imgfmt = FMT_BMP;
   char* imgout = NULL;
@@ -65,32 +78,28 @@ main(int argc, char* argv[])
   {
     if (argv[i][0] == '-')
     {
-      const char* args[NUM_ARGS] = {
-        "--nocopy", "--verbose", "--raw", "--noinline",
-        "--mask", "--vnum", "--bmp"
-      };
-      const targ_t arge[NUM_ARGS] = {
-        ARG_NOCOPY, ARG_VERBOSE, ARG_RAW, ARG_NOINLINE,
-        ARG_MASK, ARG_VNUM, ARG_BMP
-      };
       size_t j = 0;
       for (; j < NUM_ARGS; j++)
       {
-        if (!strcmp(argv[i], args[j]))
+        if (!strcmp(argv[i], args_[j]))
         {
           argcount++;
-          if (arge[j] < ARG_RESERVED)
+          if (arge_[j] < ARG_RESERVED)
           {
-            options |= arge[j];
+            options |= arge_[j];
             break;
           }
-          if (arge[j] == ARG_MASK)
+          if (arge_[j] == ARG_MASK)
           {
             mask = (uint8_t)atoi(argv[i + 1]);
           }
-          else if (arge[j] == ARG_VNUM)
+          else if (arge_[j] == ARG_VNUM)
           {
             vnum = (uint8_t)atoi(argv[i + 1]) - 1;
+          }
+          else if (arge_[j] == ARG_SCALE)
+          {
+            scale = (uint8_t)atoi(argv[i + 1]);
           }
           else // NOTE: ARG_BMP
           {
@@ -107,7 +116,7 @@ main(int argc, char* argv[])
   pdebug("finished parsing cmdln arguments");
   if (argc - argcount < NUM_MANDATORY + 1)
   {
-    return print_help_(argv[0]);
+    return phelp_(argv[0]);
   }
   if (!(options & ARG_NOCOPY))
   {
@@ -132,7 +141,7 @@ main(int argc, char* argv[])
   }
   if (imgout != NULL)
   {
-    err = qrcode_output(qr, imgfmt, imgout);
+    err = qrcode_output(qr, imgfmt, scale, imgout);
     fatalif(err);
     if (options & ARG_VERBOSE)
     {