+++ CHANGELOG.md
@@ -7,6 +7,8 @@ All notable changes to this project will be documented in this file.
 - Added option `--bmp` to write generated code to bitmap image
 - Added `pdebug` macro, which is empty if built with `NDEBUG`
 - Added option `--noinline` to suppress all inline code printing
+- Added option `-s` to scale output image resolution
+- Changed exclusive options to single `-` and single letter
 - Changed printing methods to `pbox` and `praw`
 - Changed output messages to use ANSI escape color formatting